% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{auc}
\alias{auc}
\title{Theoretical pdf of peak height}
\usage{
auc(x, u, data)
}
\arguments{
\item{x}{location of a peak}

\item{u}{a user-specified thereshold value.}

\item{data}{a numerical vector}
}
\value{
a vector of area under the curve
}
\description{
Theoretical pdf of peak height
}
\seealso{
[search_endpoint]
}
