% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak2D.R
\name{cross_area}
\alias{cross_area}
\title{Calulate cross sectional area of \eqn{f(x,y) = u}}
\usage{
cross_area(data, u)
}
\arguments{
\item{data}{a dataframe/matrix of smoothed 3D data}

\item{u}{a user-specified thereshold value}
}
\value{
a list of all cross sectional area
}
\description{
Calulate cross sectional area of \eqn{f(x,y) = u}
}
\examples{

gam <- 4
u <- 3.5
nr <- nc <- 1000   # 10000 * 10000
noise <- matrix(rnorm(nr * nc), nr, nc)
sdata <- smoothie::kernel2dsmooth(noise, "gauss", sigma=gam, nx=nr, ny=nc)
sdata <- sdata/sqrt(1/(4 * pi * gam^2))
xx <- cross_area(sdata, u=u)
xlim <- seq(0,30,0.5)
theoval <- sapply(xlim, F2x, u, gam)
plot(smooth.spline(xlim, theoval,df=5),lwd=2, type="l", ylab="Tail Probability", col="red",
ylim=c(0,1), xlab="Excursion Cross-sectional Area")
area <- ecdf.tail(xx$area$area)
lines(smooth.spline(area[,1], area[,2], df=5), lwd=2, col="black")
legend("topright", legend=c("Theoretical dist.", "Numerical dist."), col=c("red","black"), lwd=rep(2,2), bty="n")
## detect peaks by cross-sectional surface area
pvalue <- sapply(xx$area$area, F2x, u=u, gam=gam)
cluster.id <- xx$area$group[which(pvalue <= 0.05)]
lapply(cluster.id, function(i) xx$cluster[xx$cluster$group == i,])

}
\seealso{
\code{F2x}
}
