% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cross_point}
\alias{cross_point}
\title{Find the cross points of data sequence and a horizontal line}
\usage{
cross_point(data, u, n = 10000)
}
\arguments{
\item{data}{a vector}

\item{u}{a user-specified thereshold value}

\item{n}{number to data points to fit the cureve of \code{data}}
}
\value{
locations of cross points
}
\description{
Find the cross points of data sequence and a horizontal line
}
