% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak2D.R
\name{cross_volume}
\alias{cross_volume}
\title{Calculate volume above cross sectional of \eqn{f(x,y) >= u}}
\usage{
cross_volume(output.cross_area, data)
}
\arguments{
\item{output.cross_area}{output of the function \code{cross_area}}

\item{data}{Gaussian kernel smoothed data}
}
\value{
a data frame of all cross sectional volumes
}
\description{
Calculate volume above cross sectional of \eqn{f(x,y) >= u}
}
\examples{

gam <- 4
u <- 3
nr <- nc <- 1000   # 10000 * 10000
noise <- matrix(rnorm(nr * nc), nr, nc)
sdata <- smoothie::kernel2dsmooth(noise, "gauss", sigma=gam, nx=nr, ny=nc)
sdata <- sdata/sqrt(1/(4 * pi * gam^2))
xx <- cross_area(sdata, u=u)
xlim <- seq(0, 30, 0.2)
theoval <- sapply(xlim, G2v, u=u, gam)
plot(smooth.spline(xlim, theoval, df=20),lwd=2, type="l", ylab="Tail Probability", col="red",
ylim=c(0,1), xlab="Excursion Volume", main=paste("u = ", u, sep=""))
yy <- cross_volume(xx, sdata)
volume <- ecdf.tail(yy$volume)
lines(smooth.spline(volume[,1], volume[,2], df=3), lwd=2, col="black")
legend("topright", legend=c("Theoretical dist.", "Numerical dist."),
col=c("red","black"), lwd=rep(2,2), bty="n")
plotly::plot_ly(z = ~ sdata, type="surface")
#' ## detect peaks by volume above surface
pvalue <- sapply(yy$volume, G2v, u=u, gam=gam)
cluster.id <- yy$group[which(pvalue <= 0.05)]
lapply(cluster.id, function(i) xx$cluster[xx$cluster$group == i,])

}
\seealso{
\code{cross_area}
}
