% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ecdf.tail}
\alias{ecdf.tail}
\title{compute tail probability based on empirical cdf}
\usage{
ecdf.tail(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
a matrix with two columns of x and tail probability \eqn{F(x)}
}
\description{
compute tail probability based on empirical cdf
}
