\name{extremes.gui}
\alias{extremes.gui}
\title{ Open the Toolkit Main Dialog Window }
\description{
Simply opens the toolkit dialog window.
}
\usage{
extremes.gui()
}
\value{
Opens a Tcl/Tk dialog window.
}
\author{ Greg Young and Eric Gilleland }
\examples{
\dontrun{
extremes.gui()
}
}
\keyword{ misc }
