% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak2D.R
\name{F2x}
\alias{F2x}
\title{Theoretical excursion extent (cross-surface) distribution for 2D Gaussian process}
\usage{
F2x(l, u, gam = 4)
}
\arguments{
\item{l}{a value of area}

\item{u}{a user-specified thereshold value}

\item{gam}{bandwidth of Gaussian kernel}
}
\value{
right-tail probability
}
\description{
Theoretical excursion extent (cross-surface) distribution for 2D Gaussian process
}
