% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{f_x}
\alias{f_x}
\title{Theoretical pdf of peak extent}
\usage{
f_x(x, u, gam)
}
\arguments{
\item{x}{vector of quantities}

\item{u}{a user-specified thereshold value}

\item{gam}{bandwidth of Gaussian kernel}
}
\value{
density of peak extent
}
\description{
Theoretical pdf of peak extent
}
