% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak2D.R
\name{G2v}
\alias{G2v}
\title{Theoretical excursion volume (volume above cross-surface) distribution for 2D Gaussian process}
\usage{
G2v(l, u, gam = 4)
}
\arguments{
\item{l}{a value of volume}

\item{u}{a user-specified thereshold value}

\item{gam}{bandwidth of Gaussian kernel}
}
\value{
right-tail probability
}
\description{
Theoretical excursion volume (volume above cross-surface) distribution for 2D Gaussian process
}
