% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{g_v}
\alias{g_v}
\title{Theoretical pdf of peak mass}
\usage{
g_v(v, u, gam)
}
\arguments{
\item{v}{vector of quantities}

\item{u}{a user-specified thereshold value}

\item{gam}{bandwidth of Gaussian kernel}
}
\value{
density of peak mass
}
\description{
Theoretical pdf of peak mass
}
