% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{search_endpoint}
\alias{search_endpoint}
\title{Search endpoints of a peak extent}
\usage{
search_endpoint(x, u, data)
}
\arguments{
\item{x}{location of a peak}

\item{u}{a user-specified thereshold value.}

\item{data}{a numerical vector}
}
\value{
distance of a peak to its left and right endpoints
}
\description{
Search endpoints of a peak extent
}
