% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simu_peak_detect}
\alias{simu_peak_detect}
\title{Peak detection based on peak height, extent and mass}
\usage{
simu_peak_detect(
  data = NULL,
  l,
  loc,
  sigma,
  scale,
  gam = 4,
  u,
  alpha = 0.05,
  by = c("height", "extent", "mass"),
  ...
)
}
\arguments{
\item{data}{a data sequence, if not NULL, other parameters generating signal will not be used}

\item{l}{length of simulated data sequence}

\item{loc}{a vector of locations of peaks}

\item{sigma}{a vecctor of standard deviation for each peak, same length as \code{loc}}

\item{scale}{a vector of scale for each peak, same length as \code{loc}}

\item{gam}{bandwidth of gausian kernel.}

\item{u}{threshold value}

\item{alpha}{significant level for peak detection}

\item{by}{detect peaks by which, a chracter of \code{height}, \code{extent} or \code{mass}}

\item{...}{arguments in \code{rnorm}, such as mean, sd}
}
\value{
locations of significant peaks by \code{height}, \code{extent} and \code{mass}
}
\description{
Peak detection based on peak height, extent and mass
}
\examples{
loc=c(100,250,400,650,800,900)
sigma=c(4,10,15,20,10,2)
scale=c(0.4,0.5,1.2,1.8,1.2,1.2)
gam = 4
simu_peak_detect(l=1000, loc=loc, sigma=sigma, scale=scale, gam = gam, u=0.02, by="height")

}
\seealso{
[cross_point], [search_endpoint]
}
