% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simu_peak_extent}
\alias{simu_peak_extent}
\title{Gaussiann process peak extent}
\usage{
simu_peak_extent(
  n = 100,
  gam = 4,
  u,
  u.th = 3.5,
  add.extent = TRUE,
  add.density = TRUE,
  add.cdf = TRUE,
  ...
)
}
\arguments{
\item{n}{number of data points from a Gaussian distribution.}

\item{gam}{bandwidth of gausian kernel.}

\item{u}{threshold value}

\item{u.th}{threshold value for varifying the theoretical distribution of peak extent, supposed to be larger than 3}

\item{add.extent}{logical value indicates if plot the peak extent}

\item{add.density}{logical value indicates if density function should be plotted}

\item{add.cdf}{logical value indicates if plot the right tail probability distribution}

\item{...}{arguments in \code{rnorm}, such as mean, sd}
}
\value{
locations of cross points, plots of peak extent and its right-tail probability
}
\description{
Simulate the peak extent and its density of the smoothed Gaussian process
}
\examples{
simu_peak_extent(n = 200, u = 1)

}
\seealso{
[cross_point], [search_endpoint]
}
