% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simu_peak_height}
\alias{simu_peak_height}
\title{Gaussiann process peak height}
\usage{
simu_peak_height(
  n = 100,
  gamma = 4,
  add.height = TRUE,
  add.density = TRUE,
  ...
)
}
\arguments{
\item{n}{number of data points from a Gaussian distribution.}

\item{gamma}{bandwidth of gausian kernel.}

\item{add.height}{logical value indicates if peak height should be plotted}

\item{add.density}{logical value indicates if density function should be plotted}

\item{...}{arguments in \code{rnorm}, such as mean, sd}
}
\value{
location of peaks, plots of peak height and its density
}
\description{
Simulate the peak height and its density of the smoothed Gaussian process
}
\examples{
simu_peak_height(n = 100, add.height = TRUE, add.density = TRUE)

}
