\name{extRemes internal}
\alias{strip.fevd.mle}
\alias{strip.fevd.lmoments}
\alias{strip.fevd.bayesian}
\alias{consecDayMax}
\alias{cut366thDay}
\alias{densplot.evd}
\alias{eeplot}
\alias{plot.ee}
\alias{fevdPriorDefault}
\alias{fevdProposalDefault}
\alias{grlevd}
\alias{grlevdTracer}
\alias{histplot.evd}
\alias{hwmiFun}
\alias{hwmidFun}
\alias{hwyear}
\alias{initializer}
\alias{initializer.lmoments}
\alias{initializer.moms}
\alias{initializer.mle}
\alias{mvThreshold}
\alias{oevd}
\alias{oevdgen}
\alias{oevd.profpar}
\alias{percentilefun}
\alias{plot.grlevdTracer}
\alias{probprob.plot.evd}
\alias{quantilefun}
\alias{quantquant.plot.evd}
\alias{quantquant2.plot.evd}
\alias{rl.fevd}
\alias{rlgrad.fevd}
\alias{rlplot.evd}
\alias{setup.design}
\alias{setup.design.default}
\alias{setup.design.fevd}
\alias{shapePriorBeta}
\alias{will.accept}
\title{
  extRemes Internal and Secondary Functions
}

\description{
Listed below are supporting functions for the major methods in extRemes.
}

\keyword{internal}
\keyword{misc}
