\name{extremes.gui}
\alias{extremes.gui}
\title{ open the toolkit main dialog window }
\description{
Simply opens the toolkit dialog window.
}
\usage{
extremes.gui()
}
\value{
Opens a Tcl/Tk dialog window.
}
\author{ Greg Young and Eric Gilleland }
\note{ This function is called on start-up (i.e., when library( extRemes) is invoked).
This function may be used to re-open the main dialog window if it is closed while the
package is still loaded. }

\examples{
## Don't run
# extremes()
}
\keyword{ misc }% at least one, from doc/KEYWORDS
