\name{extRemes}
\alias{extRemes}
\title{extRemes -- The Extremes Toolkit: Weather and Climate Applications of Extreme Value Statistics}
\description{
'extRemes' is a graphical user interface (GUI) for modeling extreme values.  This package relies on the
'ismev' package, and most of the underlying functions are from this package.  'ismev' is an R port of
Stuart Coles' S-Plus extreme value statistical routines.

For help, see the tutorial at:

http://www.esig.ucar.edu/extremevalues/evtk.shtml
}
\keyword{misc}
