% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Fx}
\alias{Fx}
\title{Theoretical tail cdf of peak extent}
\usage{
Fx(x, u, gam)
}
\arguments{
\item{x}{vector of quantiles}

\item{u}{a user-specified thereshold value.}

\item{gam}{bandwidth of Gaussian kernel}
}
\value{
right-tail probability of peak extent
}
\description{
Theoretical tail cdf of peak extent
}
\seealso{
fx
}
