% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Gv}
\alias{Gv}
\title{Theoretical tail cdf of peak mass}
\usage{
Gv(V, u, gam)
}
\arguments{
\item{V}{vector of quantiles}

\item{u}{a user-specified thereshold value}

\item{gam}{bandwidth of Gaussian kernel}
}
\value{
right-tail probability of peak height
}
\description{
Theoretical tail cdf of peak mass
}
\seealso{
\code{g_v}
}
