% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simu_peak_detect}
\alias{simu_peak_detect}
\title{Simulate peak detection based on peak height}
\usage{
simu_peak_detect(
  data = NULL,
  l,
  loc,
  sigma,
  scale,
  gamma = 4,
  u,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{data}{a data sequence, if not NULL, other parameters generating signal will not be used}

\item{l}{length of simulated data sequence}

\item{loc}{a vector of locations of peaks}

\item{sigma}{a vecctor of standard deviation for each peak, same length as \code{loc}}

\item{scale}{a vector of scale for each peak, same length as \code{loc}}

\item{gamma}{bandwidth of gausian kernel.}

\item{u}{threshold value}

\item{alpha}{significant level for peak detection}
}
\value{
locations of cross points, plots of peak extent and its right-tail probability
}
\description{
Simulate peak detection based on peak height
}
\examples{
simu_peak_detect(l=1000, loc=c(100,250,400,650,800,900), sigma=c(4,10,15,20,10,2), scale=c(0.4,0.5,1.2,1.8,1.2,1.2), u=0.02)

}
\seealso{
[cross_point], [search_endpoint]
}
