% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simu_peak_extent}
\alias{simu_peak_extent}
\title{Find the cross points of data sequence and a horizontal line}
\usage{
simu_peak_extent(
  n = 100,
  gamma = 4,
  u,
  add.extent = TRUE,
  add.density = TRUE,
  add.cdf = TRUE,
  ...
)
}
\arguments{
\item{n}{number of data points from a Gaussian distribution.}

\item{gamma}{bandwidth of gausian kernel.}

\item{u}{threshold value}

\item{add.extent}{logical value indicates if plot the peak extent}

\item{add.density}{logical value indicates if density function should be plotted}

\item{add.cdf}{logical value indicates if plot the right tail probability distribution}
}
\value{
locations of cross points, plots of peak extent and its right-tail probability
}
\description{
Find the cross points of data sequence and a horizontal line
}
\examples{
simu_peak_extent(n = 100, u = 3.5)

}
\seealso{
[cross_point], [search_endpoint]
}
